+++ Extra/File Templates/Component.xctemplate/___FILEBASENAME___Appearance.swift
@@ -1,7 +1,0 @@
-//  ___FILEHEADER___
-
-import SwiftUI
-
-struct ___VARIABLE_productName___Appearance {
-    var backgroundColor: Color = ColorBook.background
-}