+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Core - Services
 
 - `default` in `ImageRepositoryCacheStorage` is changed from `memory` to `all`
 
+Extra - File Templates
+
+- File templates are removed
+
 ### [8.2.0(115)](https://github.com/VakhoKontridze/VCore/releases/tag/8.2.0) — 2026 06 03
 
 Core - Extensions